+++ lib/src/route_controller.dart
@@ -122,10 +122,19 @@ class RouteController {
     }
   }
 
+  _TPreservationStrategy _preservationStrategy = defaultPreservationStrategy;
+
+  static _TPreservationStrategy defaultPreservationStrategy =
+      (routeBuider, routeState) =>
+          routeBuider.shouldPreserve || routeState.shouldPreserve;
+
+  void setPreservationStrategy(_TPreservationStrategy shouldPreserve) =>
+      _preservationStrategy = shouldPreserve;
+
   void _maybeRemoveStaleRoute(RouteState routeState) {
-    final builder = _getBuilderByPath(routeState.uri);
-    if (builder == null) return;
-    if (!builder.shouldPreserve && !routeState.shouldPreserve) {
+    final routeBuilder = _getBuilderByPath(routeState.uri);
+    if (routeBuilder == null) return;
+    if (!_preservationStrategy(routeBuilder, routeState)) {
       _widgetCache[routeState] = SizedBox.shrink(key: routeState.key);
     }
   }
@@ -235,9 +244,8 @@ class RouteController {
     if (index < 0 || index >= state.routes.length) return false;
 
     _previousRouteForTransition = currentRouteState;
-    _nextAnimationEffect = index < state.index
-        ? backwardAnimationEffect
-        : forwardAnimationEffect;
+    _nextAnimationEffect =
+        index < state.index ? backwardAnimationEffect : forwardAnimationEffect;
 
     final newRoute = state.routes[index];
     addToCache([newRoute]); // Ensure widget exists before navigating.
@@ -377,9 +385,8 @@ class RouteController {
       },
       builder: (context, results) {
         final children = _widgetCache.values.toList();
-        final layerEffects = results.isNotEmpty
-            ? results.map((e) => e.data).first
-            : null;
+        final layerEffects =
+            results.isNotEmpty ? results.map((e) => e.data).first : null;
         return PrioritizedIndexedStack(
           indices: [
             _indexOfRouteState(routeState),
@@ -404,8 +411,8 @@ class RouteController {
   }
 
   static RouteController of(BuildContext context) {
-    final provider = context
-        .dependOnInheritedWidgetOfExactType<RouteControllerProvider>();
+    final provider =
+        context.dependOnInheritedWidgetOfExactType<RouteControllerProvider>();
     if (provider == null) {
       throw FlutterError('No RouteControllerProvider found in context');
     }
@@ -435,3 +442,8 @@ class _NavigationState {
     );
   }
 }
+
+typedef _TPreservationStrategy = bool Function(
+  RouteBuilder routeBuider,
+  RouteState routeState,
+);